refactor: talker.dart を debug ディレクトリへ移動
