+++ apps/dashboard/lib/core/router/router.dart
@@ -1,6 +1,6 @@
 import 'package:dashboard/core/debug/debug_screen.dart';
+import 'package:dashboard/core/debug/talker.dart';
 import 'package:dashboard/core/ui/main/main_screen.dart';
-import 'package:dashboard/core/util/talker.dart';
 import 'package:dashboard/features/account/ui/account_info_screen.dart';
 import 'package:dashboard/features/account/ui/profile_edit_screen.dart';
 import 'package:dashboard/features/account/ui/withdrawal_screen.dart';
